adds the output folder to the `.gitignore` file and minor changes
it also adds a `getfiledata()` function to the `urls.go` file that is used to retrieve data from a file given a url. a `download()` function is also added that downloads the content from the specified url using an http get request and returns it as a string. lastly, the `validateurl()` function is added to validate a given url string.